separate action from slice
